Make all sqlToJava methods return an Object (better for extendibility)
Reason I did this: suppose you want to define your own data type
ADMESerializer which convert the data into a String.
Would be good to extend the StringADMESerializer and just provide the
String/CustomDataType conversion. This is not possible if the
StringADMESerializer define "String" as return data type since you have
to return an object different from String.