Fix crash on race condition in multi-threading initialization
I need to make sure the initialization of the content provider component
is synchronized.

To do so I use a double check lock with volatile variable, which should
work according to a developer of the Android Dalvik / ART.

See comments of user "fadden" on the accepted answer here:
http://stackoverflow.com/questions/5717090/double-checked-locking-in-android/5717977

The race condition was in method registerComponent(), where the field
componentId was incremented by a second thread after the first thread
used it to add the uri to the matcher but before putting it in the
componenents map.

I made sure this can't happen anymore.