Support for content wrapping and loaders
this allow easy and quick handling of Loaders and Cursor-backed contents.

This features try to put togheter two requirements:
- Loaders should refresh themselves when data change.
- User should be able to work with java classes in place of Cursors
  (without manually converting all the time)

The ContentWrapper try to put togheter this two requirements by just
acting as a generic container for both the Cursor and the converted
object: it also expose methods for registering on content changes and
closing the wrapper/cursor.

A new utility class (ADMEWrapperUtil) can be used to easily obtain these
wrapper from Cursors by automatically converting them throw annotations.

I also created a loader (support/vanilla versions) for these objects
that already takes care of everything.

Unfurtunately I couldn't hide the wrapper to the LoaderCallback: the
callback has to know there's a wrapper around the object: this is
because the Loader API doesn't allow me to easily handle it in another
way or I couldn't find one.

I had to create a new dependency to the support-v4 library for defining
the Loader compat.

I used it with some annotation too.