Fix a nasty bug with recursive infinite loop on onContentChanged
The loader lifecycle does NOT guarantee that the deliverResult will be
always called.
Furthermore there are situation in which multiple onContentChanged are
received and cause the background task to be executed multiple times
followed by the onCanceled callback to throw the result away.

This commit fix both of this by keeping a reference of the observed
content and working with it to assure there is always a single content
observer in a give time. Then I make sure to unregister it as soon as on
content changed callback is received. The new content will be registered
upon receiving the refreshed result.